Move helpers from CiviUnitTestCase to MailingTestTrait
* `createMailing(...)`
* `deleteMailing(...)`
