+++ tests/phpunit/CiviTest/CiviUnitTestCase.php
@@ -57,6 +57,7 @@ class CiviUnitTestCase extends PHPUnit_Extensions_Database_TestCase {
   use \Civi\Test\GenericAssertionsTrait;
   use \Civi\Test\DbTestTrait;
   use \Civi\Test\ContactTestTrait;
+  use \Civi\Test\MailingTestTrait;
 
   /**
    *  Database has been initialized.
@@ -2413,37 +2414,6 @@ AND    ( TABLE_NAME LIKE 'civicrm_value_%' )
    */
   public function CiviUnitTestCase_fatalErrorHandler($message) {
     throw new Exception("{$message['message']}: {$message['code']}");
-  }
-
-  /**
-   * Helper function to create new mailing.
-   *
-   * @param array $params
-   *
-   * @return int
-   */
-  public function createMailing($params = array()) {
-    $params = array_merge(array(
-      'subject' => 'maild' . rand(),
-      'body_text' => 'bdkfhdskfhduew{domain.address}{action.optOutUrl}',
-      'name' => 'mailing name' . rand(),
-      'created_id' => 1,
-    ), $params);
-
-    $result = $this->callAPISuccess('Mailing', 'create', $params);
-    return $result['id'];
-  }
-
-  /**
-   * Helper function to delete mailing.
-   * @param $id
-   */
-  public function deleteMailing($id) {
-    $params = array(
-      'id' => $id,
-    );
-
-    $this->callAPISuccess('Mailing', 'delete', $params);
   }
 
   /**
